frontend/requirements.txt
Flask==2.0.2
google-cloud-dataproc==3.1.1 
google-cloud-storage==1.43.0 